+++ app.py
@@ -1,8 +1,6 @@
 import boto3
 import pandas as pd
-
 import streamlit as st
-import pandas as pd
 
 
 @st.cache_data
